chore(client): padroniza nomes de arquivo e pastas e os reorganiza
- Padroniza nome de arquivos e pastas
- Organiza estrutura de pastas e arquivos
