+++ src/components/select.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
 import classNames from 'classnames';
-import './style.scss';
+
+import './styles.scss';
 
 const Select = ({ name, label, children, required, ...rest }) => {
   const {
