+++ src/components/styles.scss
