+++ src/pages/client/Info/styles.scss
@@ -15,7 +15,7 @@
       width: 260px;
       background: #fff;
       border-radius: 16px;
-      box-shadow: 0 2px 8px #0000001a;
+      box-shadow: -4px 4px 12px 2px rgba(0, 0, 0, 0.25); 
       padding: 24px;
       display: flex;
       flex-direction: column;
@@ -60,7 +60,8 @@
         background: #fff;
         padding: 20px;
         border-radius: 16px;
-        box-shadow: 0 2px 8px #0000001a;        display: flex;
+        box-shadow: -4px 4px 12px 2px rgba(0, 0, 0, 0.25); 
+        display: flex;
         flex-direction: column;
         gap: 12px;
       }
