+++ src/routes/AppRoutes.jsx
@@ -1,8 +1,7 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Register from "../pages/cliente";
-import DataCustomer from "../pages/cliente/dataCustomer";
-import Layout from "../pages/cliente/layout";
+import { Register, Info, Layout } from "../pages/client";
+
 // Aqui você pode adicionar as rotas do seu projeto
 // Você também pode adicionar rotas aninhadas e, se quiser, até dividir elas em arquivos diferentes
 // Exemplo: clientRoutes.js, adminRoutes.js, etc. Ai faz a importação e exporta tudo aqui
@@ -10,9 +9,9 @@ const AppRoutes = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Layout />}>
-          <Route path="cliente" element={<Register />} />
-          <Route path="cliente/:id" element={<DataCustomer />} />
+        <Route path="/cliente" element={<Layout />}>
+          <Route index element={<Register />} />
+          <Route path=":id" element={<Info />} />
         </Route>
       </Routes>
     </BrowserRouter>
